requirements.txt
streamlit
google-genai
pandas
fpdf
gspread
oauth2client
streamlit-option-menu
plotly
PyPDF2
requests